Fix: Squeeze extra dimension from DataLoader output
The DataLoader was adding an extra dimension of size 1 to the batch,
resulting in an input shape of [batch_size, 1, channels, height, width]
instead of the expected [batch_size, channels, height, width]. This
caused a RuntimeError in the conv2d layer.

This commit adds a `squeeze(1)` call after loading the data from the
DataLoader, removing the extra dimension and ensuring the correct input
shape for the network. This fixes the RuntimeError and allows training
to proceed.